+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 ## 0.13.0
 
 - Added functionality to manage authorised client applications for exposed api permissions.
-- Amended the display item for exposed api permissions to show scope value instead of admin display name as it's unique.
 - Added functionality to display deleted app roles and scopes when referenced by an application with an error indicator.
+- Added a disabled label to the tooltip of disabled app roles and scopes.
+- Amended the display item for exposed api permissions to show scope value instead of admin display name as it's unique.
+- Amended the display item for app roles to show value instead of display name as it's unique.
 - Fixed an issue where displaying deleted api permissions would cause an error.
 - Fixed an issue where displaying scopes from a deleted application would cause an error.